frontend inventory listings

+++ src/main/java/org/bakerydb/backend/modelutils/InventoryUtil.java
@@ -1,8 +1,8 @@
-package org.bakerydb.backend.itemutils;
+package org.bakerydb.backend.modelutils;
 
 import org.bakerydb.util.*;
 import org.bakerydb.backend.DBConnection;
-import org.bakerydb.backend.items.InventoryItem;
+import org.bakerydb.backend.models.InventoryItem;
 import java.util.ArrayList;
 import java.sql.*;
 
@@ -15,30 +15,35 @@ public class InventoryUtil {
 
     /**
      * Add a new item to the inventory
-     * @param name the name of the item
-     * @param quantity the quantity of the item in stock
-     * @param unit the unit associated with the quatity
-     * @param reorderLevel the level at which the item should be reordered
-     * @return a Result containing either a Void or an error message
+    * @param item the InventoryItem to add
+     * @return a Result containing either the resulting key or an error message
      */
-    public Result<Void> newItem(
-        String name,
-        Integer quantity,
-        String unit,
-        Integer reorderLevel
-    ) {
+    public Result<Integer> newItem(InventoryItem item) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "INSERT INTO Inventory (name, quantity, unit, reorderLevel) VALUES (?, ?, ?, ?)";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
-            stmt.setString(1, name);
-            stmt.setInt(2, quantity);
-            stmt.setString(3, unit);
-            stmt.setInt(4, reorderLevel);
-            stmt.executeUpdate();
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query, PreparedStatement.RETURN_GENERATED_KEYS);
+            stmt.setString(1, item.getName());
+            stmt.setFloat(2, item.getQuantity());
+            stmt.setString(3, item.getUnit());
+            stmt.setFloat(4, item.getReorderLevel());
+
+            int affectedRows = stmt.executeUpdate();
+            if (affectedRows > 0) {
+                ResultSet keys = stmt.getGeneratedKeys();
+                if (keys.next())
+                    return Result.ok(keys.getInt(1));
+                else
+                    return Result.err(ErrorMessage.NO_KEY);
+            } else {
+                return Result.err(ErrorMessage.NO_CHANGE);
+            }
+            
         } catch (SQLException e) {
             return Result.err(e.toString());
         }
-        return Result.ok();
     }
 
     /**
@@ -47,9 +52,12 @@ public class InventoryUtil {
      * @return a Result containing either a Void or an error message
      */
     public Result<Void> deleteItem(Integer invId) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "DELETE FROM Inventory WHERE invId = ?";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
             stmt.setInt(1, invId);
             stmt.executeUpdate();
         } catch (SQLException e) {
@@ -64,11 +72,14 @@ public class InventoryUtil {
      * @param quantity the new quantity of the item
      * @return a Result containing either a Void or an error message
      */
-    public Result<Void> setQuantity(Integer invId, Integer quantity) {
+    public Result<Void> setQuantity(Integer invId, Float quantity) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "UPDATE Inventory SET quantity = ? WHERE invId = ?";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
-            stmt.setInt(1, quantity);
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
+            stmt.setFloat(1, quantity);
             stmt.setInt(2, invId);
             stmt.executeUpdate();
         } catch (SQLException e) {
@@ -83,12 +94,39 @@ public class InventoryUtil {
      * @param amount the amount to change the quantity by
      * @return a Result containing either a Void or an error message
      */
-    public Result<Void> changeQuantityBy(Integer invId, Integer amount) {
+    public Result<Void> changeQuantityBy(Integer invId, Float amount) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "UPDATE Inventory SET quantity = quantity + ? WHERE invId = ?";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
-            stmt.setInt(1, amount);
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
+            stmt.setFloat(1, amount);
             stmt.setInt(2, invId);
+            stmt.executeUpdate();
+        } catch (SQLException e) {
+            return Result.err(e.toString());
+        }
+        return Result.ok();
+    }
+
+    /**
+     * Update an item in the inventory
+     * @param item the InventoryItem to update
+     * @return a Result containing either a Void or an error message
+     */
+    public Result<Void> update(InventoryItem item) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
+        String query = "UPDATE Inventory SET name = ?, quantity = ?, unit = ?, reorderLevel = ? WHERE invId = ?";
+        try {
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
+            stmt.setString(1, item.getName());
+            stmt.setFloat(2, item.getQuantity());
+            stmt.setString(3, item.getUnit());
+            stmt.setFloat(4, item.getReorderLevel());
+            stmt.setInt(5, item.getInvId());
             stmt.executeUpdate();
         } catch (SQLException e) {
             return Result.err(e.toString());
@@ -101,9 +139,12 @@ public class InventoryUtil {
      * @return a Result containing either an ArrayList of Inventory items or an error message
      */
     public Result<ArrayList<InventoryItem>> fetchAll() {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "SELECT * FROM Inventory";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
             return Result.ok(InventoryItem.list(stmt.executeQuery()));
         } catch (SQLException e) {
             return Result.err(e.toString());
@@ -116,9 +157,12 @@ public class InventoryUtil {
      * @return a Result containing either an Inventory item or an error message
      */
     public Result<InventoryItem> fetch(Integer invId) {
+        if (!this.DB.isConnected())
+            return Result.err(ErrorMessage.NO_CONNECTION);
+
         String query = "SELECT * FROM Inventory WHERE invId = ?";
         try {
-            PreparedStatement stmt = DB.connection.prepareStatement(query);
+            PreparedStatement stmt = this.DB.connection.prepareStatement(query);
             stmt.setInt(1, invId);
             ResultSet result = stmt.executeQuery();
             if (result.next())
